Make `Leibniz` data type definition consistent.
