--- base/src/main/scala/Leibniz.scala
@@ -1,16 +1,0 @@
-package scalaz
-
-import data.Identity.Id
-
-object Leibniz {
-  trait ===[A, B] {
-    def subst[F[_]](fa: F[A]): F[B]
-
-    def apply(a: A): B = subst[Id](a)
-    def onF[X](fa: X => A): X => B = subst[X => ?](fa)
-  }
-
-  def refl[A]: A === A = new (A === A) {
-    def subst[F[_]](fa: F[A]): F[A] = fa
-  }
-}
